github-classroom/template-week03-docker-opensearch/requirements.txt
pytest>=7.0.0
pyyaml>=6.0
opensearch-py>=2.0.0
requests>=2.28.0
